requirements.txt
torch==1.9
torchvision
numpy
scipy
scikit-image
pillow
pandas
tqdm
dominate
